Add compiler check to make MinGW work fine.

+++ setup_libuv.py
@@ -1,3 +1,4 @@
+
 import errno
 import os
 import shutil
@@ -82,16 +83,16 @@ class libuv_build_ext(build_ext):
         elif sys.platform == 'darwin':
             self.extensions[0].extra_link_args = ['-framework', 'CoreServices']
         elif sys.platform == 'win32':
-            self.library_dirs.append(self.libuv_dir+r'\release\lib')
+            self.library_dirs.append(os.path.join(self.libuv_dir, 'Release', 'lib'))
             self.libraries.append('iphlpapi')
             self.libraries.append('psapi')
             self.libraries.append('ws2_32')
-            self.libraries.append('kernel32')
-            self.libraries.append('user32')
             self.libraries.append('advapi32')
 
     def get_libuv(self):
         #self.debug_mode =  bool(self.debug) or hasattr(sys, 'gettotalrefcount')
+        self.cc = self.compiler.compiler_type
+        
         def download_libuv():
             log.info('Downloading libuv...')
             makedirs(self.libuv_dir)
@@ -105,7 +106,7 @@ class libuv_build_ext(build_ext):
             cflags = '-fPIC'
             env = os.environ.copy()
             env['CFLAGS'] = ' '.join(x for x in (cflags, env.get('CFLAGS', None)) if x)
-            if sys.platform=="win32":
+            if sys.platform=='win32' and self.cc=='msvc':
                 log.info('build libuv on win32...')
                 os.system(self.libuv_dir + r"\vcbuild.bat release")
                 #exec_process(['cmd.exe','/C','vcbuild.bat release'], cwd=self.libuv_dir, env=env)
@@ -121,19 +122,19 @@ class libuv_build_ext(build_ext):
             build_libuv()
         else:
             if self.libuv_clean_compile:
-                if sys.platform=='win32':
+                if sys.platform=="win32" and self.cc=="msvc":
                     os.system('rd /s /q' + os.path.join(self.libuv_dir, 'Release'))
                 else:
                     exec_process(['make', 'clean'], cwd=self.libuv_dir)
                     
-            if sys.platform!='win32':
-                if not os.path.exists(os.path.join(self.libuv_dir, 'libuv.a')):
+            if sys.platform=="win32" and self.cc=="msvc":
+                if not os.path.exists(os.path.join(self.libuv_dir, 'Release', 'lib', 'uv.lib')):
                     log.info('libuv needs to be compiled.')
                     build_libuv()
                 else:
                     log.info('No need to build libuv.')  
             else:
-                if not os.path.exists(os.path.join(self.libuv_dir, 'Release', 'lib', 'uv.lib')):
+                if not os.path.exists(os.path.join(self.libuv_dir, 'libuv.a')):
                     log.info('libuv needs to be compiled.')
                     build_libuv()
                 else:
@@ -147,6 +148,7 @@ class libuv_sdist(sdist):
     libuv_patches  = libuv_build_ext.libuv_patches
 
     def initialize_options(self):
+        log.info('libuv_sdist.initialize_options')
         sdist.initialize_options(self)
         rmtree('deps')
         makedirs(self.libuv_dir)
